core: delete the shim APIs that no longer vary by Spark version
Fifteen of them: the callers read the underlying Spark API directly, and the
two eval-mode predicates plus the file-metadata helper move to gluten-substrait
utils. createParquetFilters and widerDecimalType stay, since LegacyBehaviorPolicy
and DecimalPrecision are named differently on 3.4 than on 3.5+.
