+++ gluten-substrait/src/main/scala/org/apache/gluten/execution/ScanTransformerFactory.scala
@@ -44,9 +44,8 @@ object ScanTransformerFactory {
       batchScanExec.output,
       batchScanExec.scan,
       batchScanExec.runtimeFilters,
-      keyGroupedPartitioning =
-        SparkShimLoader.getSparkShims.getKeyGroupedPartitioning(batchScanExec),
-      table = SparkShimLoader.getSparkShims.getBatchScanExecTable(batchScanExec)
+      keyGroupedPartitioning = batchScanExec.keyGroupedPartitioning,
+      table = batchScanExec.table
     )
   }
 }
