+++ gluten-substrait/src/main/scala/org/apache/gluten/extension/columnar/offload/OffloadSingleNodeRules.scala
@@ -183,6 +183,16 @@ case class OffloadOthers() extends OffloadSingleNode with LogLevelUtil {
 }
 
 object OffloadOthers {
+  // A limit of -1 means only an offset was given, so fetch every remaining row.
+  private def limitAndOffset(limit: Int, offset: Int): (Int, Int) = {
+    if (limit == -1) {
+      (Int.MaxValue, offset)
+    } else {
+      assert(limit > offset)
+      (limit - offset, offset)
+    }
+  }
+
   // Utility to replace single node within transformed Gluten node.
   // Children will be preserved as they are as children of the output node.
   //
@@ -243,7 +253,7 @@ object OffloadOthers {
           SortExecTransformer(plan.sortOrder, plan.global, child, plan.testSpillFrequency)
         case plan: TakeOrderedAndProjectExec =>
           val child = plan.child
-          val (limit, offset) = SparkShimLoader.getSparkShims.getLimitAndOffsetFromTopK(plan)
+          val (limit, offset) = limitAndOffset(plan.limit, plan.offset)
           TakeOrderedAndProjectExecTransformer(
             limit,
             plan.sortOrder,
@@ -269,8 +279,7 @@ object OffloadOthers {
           )
         case plan: GlobalLimitExec =>
           val child = plan.child
-          val (limit, offset) =
-            SparkShimLoader.getSparkShims.getLimitAndOffsetFromGlobalLimit(plan)
+          val (limit, offset) = limitAndOffset(plan.limit, plan.offset)
           LimitExecTransformer(child, offset, limit)
         case plan: LocalLimitExec =>
           val child = plan.child
